feat: made has_idx_mutated() notice function
